Move PartialSQLRouteExecutor logic to TablelessRouteEngine to handle tableless sql route
