--- infra/route/src/main/java/org/apache/shardingsphere/infra/route/engine/SQLRouteExecutorDecider.java
@@ -1,36 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.infra.route.engine;
-
-import org.apache.shardingsphere.infra.spi.annotation.SingletonSPI;
-import org.apache.shardingsphere.sql.parser.statement.core.statement.SQLStatement;
-
-/**
- * SQL route executor decider.
- */
-@SingletonSPI
-public interface SQLRouteExecutorDecider {
-    
-    /**
-     * Judge is need route all resources.
-     *
-     * @param sqlStatement SQL statement
-     * @return is need route all resources or not
-     */
-    boolean isNeedRouteAll(SQLStatement sqlStatement);
-}
