+++ infra/route/src/main/java/org/apache/shardingsphere/infra/route/engine/tableless/TablelessRouteEngine.java
@@ -15,29 +15,23 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.route.engine.type;
+package org.apache.shardingsphere.infra.route.engine.tableless;
 
 import org.apache.shardingsphere.infra.metadata.database.ShardingSphereDatabase;
 import org.apache.shardingsphere.infra.metadata.database.rule.RuleMetaData;
 import org.apache.shardingsphere.infra.route.context.RouteContext;
-import org.apache.shardingsphere.infra.route.context.RouteMapper;
-import org.apache.shardingsphere.infra.route.context.RouteUnit;
-import org.apache.shardingsphere.infra.route.engine.SQLRouteExecutor;
-import org.apache.shardingsphere.infra.session.query.QueryContext;
-
-import java.util.Collections;
 
 /**
- * All SQL route executor.
+ * Tableless route engine.
  */
-public final class AllSQLRouteExecutor implements SQLRouteExecutor {
+public interface TablelessRouteEngine {
     
-    @Override
-    public RouteContext route(final QueryContext queryContext, final RuleMetaData globalRuleMetaData, final ShardingSphereDatabase database) {
-        RouteContext result = new RouteContext();
-        for (String each : database.getResourceMetaData().getStorageUnits().keySet()) {
-            result.getRouteUnits().add(new RouteUnit(new RouteMapper(each, each), Collections.emptyList()));
-        }
-        return result;
-    }
+    /**
+     * Route.
+     *
+     * @param globalRuleMetaData global rule meta data
+     * @param database database
+     * @return route context
+     */
+    RouteContext route(RuleMetaData globalRuleMetaData, ShardingSphereDatabase database);
 }
